Move tag consts to src/consts/tags
