+++ src/components/sidebar-followed-tag.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router';
 
 import { truncate } from 'grapheme-utils';
 import TagIcon from './tag-icon';
-import { TAG_HASHTAG, TAG_SCHOOL, TAG_LOCATION } from '../utils/tags';
+import { TAG_HASHTAG, TAG_SCHOOL, TAG_LOCATION } from '../consts/tags';
 
 export default class SidebarFollowedTag extends React.Component {
   static displayName = 'SidebarFollowedTag';
