+++ src/consts/tags.js
@@ -1,0 +1,6 @@
+export const TAG_HASHTAG = 'TAG_HASHTAG';
+export const TAG_SCHOOL = 'TAG_SCHOOL';
+export const TAG_MENTION = 'TAG_MENTION';
+export const TAG_LOCATION = 'TAG_LOCATION';
+export const TAG_EVENT = 'TAG_EVENT';
+export const IMPLEMENTED_TAGS = [TAG_HASHTAG, TAG_LOCATION, TAG_SCHOOL];
