+++ src/utils/tags.js
@@ -1,8 +1,5 @@
-export const TAG_HASHTAG = 'TAG_HASHTAG';
-export const TAG_SCHOOL = 'TAG_SCHOOL';
-export const TAG_MENTION = 'TAG_MENTION';
-export const TAG_LOCATION = 'TAG_LOCATION';
-export const TAG_EVENT = 'TAG_EVENT';
+import { TAG_HASHTAG, TAG_LOCATION, TAG_SCHOOL } from '../consts/tags';
+
 
 /**
  * Converts hashtags(labels), schools, and other tags to the same format.
